melanomadiagnosis/requirements.txt
Flask
tensorflow
Pillow
numpy
flask-cors
groq
